requirements.txt
fastapi
uvicorn
psycopg2-binary
hvac
python-dotenv
pydantic
